requirements.txt
gym==0.10.5
numpy>=1.19.5
matplotlib>=3.3.2
pyglet>=1.5.0
scikit-learn>=0.23.2
shapely>=1.7.1
simplekml>=1.3.2
torch>=1.6.0
wadl-planner==1.0.1
ortools>=8.2
